+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [unreleased]
 
+### Added
+
+- `meci` parser for TeraChem. Grabs upper and lower energies from minimal energy conical intersection optimizations and also the geometry at each step.
+
 ## [0.6.2] - 2024-08-13
 
 ### Added